Sync InteractionMessage definitions
